requirements.txt
Flask
PyJWT
peewee
bcrypt
requests
